+++ frontend_forms/static/frontend_forms/js/frontend_forms.jsx
@@ -872,6 +872,52 @@ window.FrontendForms = (function() {
         }
     }
 
+    /* Apply multiSelect with search capabilities to a '<select ... multiple="">' element */
+
+    /*
+     * TODO: add support for SELECT ALL
+     *
+     * see: https://stackoverflow.com/questions/39007371/jquery-multiselect-select-all-and-with-filtered-search
+     */
+
+    function apply_multiselect(elements) {
+        $(elements).multiSelect({
+            selectableHeader: "<input type='text' class='form-control search-input' autocomplete='off' placeholder='cerca ...'>",
+            selectionHeader: "<input type='text' class='form-control search-input' autocomplete='off' placeholder='cerca ...'>",
+            afterInit: function(ms) {
+                var that = this,
+                    $selectableSearch = that.$selectableUl.prev(),
+                    $selectionSearch = that.$selectionUl.prev(),
+                    selectableSearchString = '#' + that.$container.attr('id') + ' .ms-elem-selectable:not(.ms-selected)',
+                    selectionSearchString = '#' + that.$container.attr('id') + ' .ms-elem-selection.ms-selected';
+
+                that.qs1 = $selectableSearch.quicksearch(selectableSearchString)
+                    .on('keydown', function(e) {
+                        if (e.which === 40) {
+                            that.$selectableUl.focus();
+                            return false;
+                        }
+                    });
+
+                that.qs2 = $selectionSearch.quicksearch(selectionSearchString)
+                    .on('keydown', function(e) {
+                        if (e.which == 40) {
+                            that.$selectionUl.focus();
+                            return false;
+                        }
+                    });
+            },
+            afterSelect: function() {
+                this.qs1.cache();
+                this.qs2.cache();
+            },
+            afterDeselect: function() {
+                this.qs1.cache();
+                this.qs2.cache();
+            }
+        });
+    }
+
     return {
         display_server_error: display_server_error,
         redirect: redirect,
@@ -890,7 +936,8 @@ window.FrontendForms = (function() {
         confirmRemoteAction: confirmRemoteAction,
         downloadFromAjaxPost: downloadFromAjaxPost,
         querystring_parse: querystring_parse,
-        set_datepicker_defaults: set_datepicker_defaults
+        set_datepicker_defaults: set_datepicker_defaults,
+        apply_multiselect: apply_multiselect
     };
 
 })();